requirements.txt
Flask==3.0.2
praw==7.7.1
transformers==4.41.2
torch==2.3.0
nltk==3.8.1
emoji==2.11.1
langdetect==1.0.9
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
python-dotenv==1.0.0